feat: community layout & board page (#4127)
